Improve organization, naming, and documentation
